Adding final MC sprite and 2 test sprites for NPC 1
The final MC sprite is the Nervous sprite. The test sprites are for NPC 1 to figure out if zooming into the full body for the VN part or making new sprites for the VN part is better. The full body of the NPC is a clickable interactable that triggers the VN section w/ their dialogue.