First area Narrative
I redid the narrative removing the DETOUR Command and replacing it with the JUMP command. It will need to be replaced with the point and click mechanic since it's not actual branching dialogue, but I figured NQ and I would go over that post alpha presentation. It should be identical to the sample script. I got 4 minor errors in that it was unable to find certain Nisaea expressions and the townsquare location when trying to find it in unity, but I figure that was because it hadn't been officially added by Art yet and would be tomorrow. Tested in Yarnspinner web and it worked (I still cant access anything in unity on my mac however).