Made scripts for intro and outro
Pasted in writing for intro area, set up outro nodes, connected the intro and outro node scene changes for the previous nodes where applicable. Haven't fiddled with the if else statements yet or pasted in Outro text, but will work on that today, just wanted to do a PR so the assets for the storybook could start being placed in the Intro area if folks were ready for that.